Fix overwriting globals and make buddy_alloc work with size_ts

+++ include/buddy_alloc.h
@@ -1,9 +1,9 @@
 #include <stdbool.h>
 #include <stdlib.h>
 
-void heap_init(void* mem_pool, int mem_size);
-void *heap_alloc(int size);
+void heap_init(void* mem_pool, size_t mem_size);
+void *heap_alloc(size_t size);
 bool heap_free(void *blk);
-int heap_done();
+size_t heap_done();
 void* heap_realloc(void* blk, size_t new_size);
 void* heap_calloc(size_t cnt, size_t size);
